Split the partitioning code
Move files of the partitioning modules to new directories.
