+++ pyanaconda/modules/storage/partitioning/automatic/automatic_interface.py
