+++ pyanaconda/modules/storage/partitioning/automatic/noninteractive_partitioning.py
