+++ pyanaconda/modules/storage/partitioning/automatic/utils.py
