+++ pyanaconda/modules/storage/partitioning/custom/Makefile.am
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2018  Red Hat, Inc.
+# Copyright (C) 2019  Red Hat, Inc.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU Lesser General Public License as published
@@ -14,10 +14,8 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-SUBDIRS = automatic blivet custom interactive manual
-
 pkgpyexecdir = $(pyexecdir)/py$(PACKAGE_NAME)
-partitioning_moduledir = $(pkgpyexecdir)/modules/storage/partitioning
-partitioning_module_PYTHON = $(srcdir)/*.py
+custom_moduledir = $(pkgpyexecdir)/modules/storage/partitioning/custom
+custom_module_PYTHON = $(srcdir)/*.py
 
 MAINTAINERCLEANFILES = Makefile.in
