+++ pyanaconda/modules/storage/partitioning/interactive/interactive_interface.py
