+++ pyanaconda/modules/storage/partitioning/interactive/utils.py
