+++ pyanaconda/modules/storage/partitioning/manual/manual_interface.py
