+++ pyanaconda/modules/storage/partitioning/manual/manual_module.py
@@ -26,8 +26,10 @@ from pyanaconda.modules.common.constants.objects import MANUAL_PARTITIONING
 from pyanaconda.modules.common.structures.partitioning import MountPointRequest
 from pyanaconda.modules.storage.partitioning.base import PartitioningModule
 from pyanaconda.modules.storage.partitioning.constants import PartitioningMethod
-from pyanaconda.modules.storage.partitioning.manual_interface import ManualPartitioningInterface
-from pyanaconda.modules.storage.partitioning.manual_partitioning import ManualPartitioningTask
+from pyanaconda.modules.storage.partitioning.manual.manual_interface import \
+    ManualPartitioningInterface
+from pyanaconda.modules.storage.partitioning.manual.manual_partitioning import \
+    ManualPartitioningTask
 
 log = get_module_logger(__name__)
