+++ pyanaconda/modules/storage/partitioning/manual/manual_partitioning.py
@@ -20,7 +20,7 @@ from pykickstart.errors import KickstartParseError
 
 from pyanaconda.anaconda_loggers import get_module_logger
 from pyanaconda.core.i18n import _
-from pyanaconda.modules.storage.partitioning.noninteractive_partitioning import \
+from pyanaconda.modules.storage.partitioning.automatic.noninteractive_partitioning import \
     NonInteractivePartitioningTask
 
 log = get_module_logger(__name__)
